requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
django==4.2.2
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
pytz==2024.1
    # via djangorestframework
sqlparse==0.4.4
    # via django
typing-extensions==4.10.0
    # via asgiref
tzdata==2024.1
    # via django
